Refactor/#59 : 파일 형식 변경
