+++ server/gateway/src/app.module.ts
@@ -5,7 +5,7 @@ import { AuthModule } from './auth/auth.module';
 import { TicketModule } from './ticket/ticket.module';
 import { CoreModule } from './core/core.module';
 import { SocketModule } from './socket/socket.module';
-import { LoggerMiddleware } from './middleware/logger.middleware';
+import { LoggerMiddleware } from './common/middleware/logger.middleware';
 
 @Module({
   imports: [AuthModule, TicketModule, CoreModule, SocketModule],
