+++ server/gateway/src/common/middleware/logger.middleware.ts
