+++ server/gateway/src/common/types/custom.response.d.ts
