+++ server/gateway/src/common/types/index.ts
