+++ server/gateway/test/fanup.e2e-spec.ts
@@ -5,7 +5,7 @@
 import { Test } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import { Socket as ClientSocket, io } from 'socket.io-client';
-import { ChatGateway } from '../../src/socket/chat.gateway';
+import { FanUPGateway } from '../src/socket/fanup/fanup.gateway';
 
 async function createNestApp(...gateways): Promise<INestApplication> {
   const testingModule = await Test.createTestingModule({
@@ -24,7 +24,9 @@ describe('FanUP 소켓 테스트', () => {
     socketID: '',
     email: 'jinsung1048@gmail.com',
     roomName: 'boostcamp',
-    message: 'messi',
+    offer: 'client1_offer',
+    answer: 'client1_answer',
+    ice: 'client1_ice',
   };
 
   let client2: ClientSocket;
@@ -32,22 +34,24 @@ describe('FanUP 소켓 테스트', () => {
     socketID: '',
     email: 'jinsung1048@nate.com',
     roomName: 'boostcamp',
-    message: 'ronaldo',
+    offer: 'client2_offer',
+    answer: 'client2_answer',
+    ice: 'client2_ice',
   };
 
   let client3: ClientSocket;
   let client3SocketID: string;
 
   beforeAll(async () => {
-    app = await createNestApp(ChatGateway);
+    app = await createNestApp(FanUPGateway);
     await app.listen(3000);
 
-    client = io('http://localhost:3000/socket/chat');
-    client1 = io('http://localhost:3000/socket/chat');
-    client2 = io('http://localhost:3000/socket/chat');
-    client3 = io('http://localhost:3000/socket/chat');
+    client = io('http://localhost:3000/socket/fanup');
+    client1 = io('http://localhost:3000/socket/fanup');
+    client2 = io('http://localhost:3000/socket/fanup');
+    client3 = io('http://localhost:3000/socket/fanup');
 
-    client1.emit('join-chat-room', {
+    client1.emit('join_room', {
       email: client1Data.email,
       roomName: client1Data.roomName,
     });
@@ -58,7 +62,7 @@ describe('FanUP 소켓 테스트', () => {
       });
     });
 
-    client2.emit('join-chat-room', {
+    client2.emit('join_room', {
       email: client2Data.email,
       roomName: client2Data.roomName,
     });
@@ -69,9 +73,9 @@ describe('FanUP 소켓 테스트', () => {
       });
     });
 
-    client3.emit('join-chat-room', {
+    client3.emit('join_room', {
       email: 'disconnected_email',
-      roomName: client1Data.roomName,
+      roomName: client2Data.roomName,
     });
 
     client3SocketID = await new Promise<string>((resolve) => {
@@ -88,10 +92,10 @@ describe('FanUP 소켓 테스트', () => {
     await app.close();
   }, 10000);
 
-  it('join-chat-room 테스트', async () => {
-    client.emit('join-chat-room', {
+  it('join_room 테스트', async () => {
+    client.emit('join_room', {
       email: 'client@email',
-      roomName: 'join-chat-room',
+      roomName: 'join_test',
     });
 
     await new Promise<void>((resolve) => {
@@ -102,19 +106,52 @@ describe('FanUP 소켓 테스트', () => {
     });
   }, 10000);
 
-  it('send-message 테스트', async () => {
-    client1.emit('send-message', {
+  it('offer 테스트', async () => {
+    client1.emit('offer', {
       email: client1Data.email,
-      roomName: client1Data.roomName,
-      isArtist: true,
-      message: client1Data.message,
+      offer: client1Data.offer,
+      targetSocketID: client2Data.socketID,
     });
 
     await new Promise<void>((resolve) => {
-      client2.on('receive-message', (data) => {
+      client2.on('offer', (data) => {
         expect(data.email).toEqual(client1Data.email);
-        expect(data.isArtist).toEqual(true);
-        expect(data.message).toEqual(client1Data.message);
+        expect(data.offer).toEqual(client1Data.offer);
+        expect(data.socketID).toEqual(client1Data.socketID);
+        resolve();
+      });
+    });
+  }, 10000);
+
+  it('answer 테스트', async () => {
+    client1.emit('answer', {
+      email: client1Data.email,
+      answer: client1Data.answer,
+      targetSocketID: client2Data.socketID,
+    });
+
+    await new Promise<void>((resolve) => {
+      client2.on('answer', (data) => {
+        expect(data.email).toEqual(client1Data.email);
+        expect(data.answer).toEqual(client1Data.answer);
+        expect(data.socketID).toEqual(client1Data.socketID);
+        resolve();
+      });
+    });
+  }, 10000);
+
+  it('ice 테스트', async () => {
+    client1.emit('ice', {
+      email: client1Data.email,
+      ice: client1Data.ice,
+      targetSocketID: client2Data.socketID,
+    });
+
+    await new Promise<void>((resolve) => {
+      client2.on('ice', (data) => {
+        expect(data.email).toEqual(client1Data.email);
+        expect(data.ice).toEqual(client1Data.ice);
+        expect(data.socketID).toEqual(client1Data.socketID);
         resolve();
       });
     });
